requirements.txt
deepgram-sdk
groq
python-dotenv
requests
flask